requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
sqlparse==0.5.3
drf-yasg==1.21.8
django-cors-headers==4.6.0